refactor(vats): rename the core bootstrap modules to src/core
