+++ packages/vats/src/core/types.js
@@ -1,0 +1,53 @@
+// @ts-check
+
+/** @typedef { import('@agoric/eventual-send').EProxy } EProxy */
+
+/**
+ * SwingSet types
+ *
+ * @typedef { ERef<ReturnType<typeof
+ *   import('@agoric/swingset-vat/src/devices/mailbox-src.js').buildRootDeviceNode>> } MailboxDevice
+ * @typedef { ERef<ReturnType<typeof
+ *   import('@agoric/swingset-vat/src/vats/vat-tp.js').buildRootObject>> } VattpVat
+ * @typedef { ERef<ReturnType<typeof
+ *   import('@agoric/swingset-vat/src/kernel/vatAdmin/vatAdminWrapper.js').buildRootObject>> } VatAdminVat
+ * @typedef { ERef<ReturnType<typeof
+ *   import('@agoric/swingset-vat/src/vats/vat-timerWrapper.js').buildRootObject>> } TimerVat
+ * @typedef {{ getChainBundle: () => unknown }} ChainBundler
+ *
+ * See deliverToController in packages/SwingSet/src/vats/comms/controller.js
+ * @typedef {ERef<{
+ *   addRemote: (name: string, tx: unknown, rx: unknown) => void,
+ *   addEgress: (addr: string, ix: number, provider: unknown) => ERef<ChainBundler>,
+ * }>} CommsVatRoot
+ *
+ * @typedef {{
+ *   comms: CommsVatRoot,
+ *   timer: TimerVat,
+ *   vatAdmin: VatAdminVat,
+ *   vattp: VattpVat,
+ * }} SwingsetVats
+ * @typedef {{
+ *   mailbox: MailboxDevice,
+ *   vatAdmin: unknown,
+ * }} SwingsetDevices
+ */
+
+/**
+ * @typedef {{ resolve: (v: T) => void }} Producer<T>
+ * @template T
+ */
+/**
+ * @typedef {(name: string) => T} VatLoader<T>
+ * @template T
+ */
+/**
+ * @typedef {{
+ *   consume: Record<string, Promise<unknown>>,
+ *   produce: Record<string, Producer<unknown>>,
+ * }} PromiseSpace
+ *
+ * @typedef {{
+ *   assignBundle: (ps: Record<string, (addr: string) => void>) => void
+ * }} ClientConfig
+ */
